backend/requirements.txt
# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0


# Validation & settings
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# HTTP client
httpx>=0.27.0

# Queue
redis>=5.0.0
rq>=1.16.0

# Rate limiting
slowapi>=0.1.9

# Scheduling
apscheduler>=3.10.0

# Web scraping / crawling
markdownify>=0.12.0
firecrawl-py>=0.0.16

# Utilities
python-multipart>=0.0.9
groq>=0.9.0
